+++ Utility基础类大全/条形码帮助类/BarCodeHelper.cs
